+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ## 1.1 branch
 ### 1.1.1
 * updated Font Awesome to 4.3.0 version;
+* added some methods to `Plugin` and `System` utilities;
 * added the changelog file.